check userId before delete and update

+++ handlers/drives.py
@@ -11,14 +11,19 @@ logger = logging.getLogger('rishacar.' + __name__)
 class DrivesHandler(BaseHandler):
   async def get(self, id):
     driveFunc = DriveFunctions()
+    # get a drive by id
     if id:
       result = await driveFunc.getDrive(id)
+      result['_id'] = str(result['_id'])
+      self.write(json_util.dumps(result))
+      self.finish()
+    # get filtered rides
     elif self.request.body:
       data = json.loads(self.request.body)
       result = await driveFunc.getFilteredDrives(data)
+    # get all rides
     else:
       result = await driveFunc.getDrives()
-
     if result:
       self.write(json_util.dumps(super(DrivesHandler, self).listToDict(result)))
     else:
@@ -29,14 +34,18 @@ class DrivesHandler(BaseHandler):
   async def post(self, _):
     if self.request.body:
       data = json.loads(self.request.body)
-      driveFunc = DriveFunctions()
-      result = await driveFunc.insertDrive(data)
-      print(result)
-      if result:
-        self.write({"message":"drive added successfully"})
+      if self.verify_data(data):
+        driveFunc = DriveFunctions()
+        result = await driveFunc.insertDrive(data)
+        if result:
+          self.write({"message":"drive added successfully"})
+        else:
+          self.set_status(500)
+          self.write({"message":"database error"})
+          self.finish()
       else:
-        self.set_status(500)
-        self.write({"message":"database error"})
+        self.set_status(400)
+        self.write({"message":"missing some data"})
         self.finish()
     else:
       self.set_status(400)
@@ -45,31 +54,31 @@ class DrivesHandler(BaseHandler):
 
   async def delete(self, id):
     if id:
-      driveFunc = DriveFunctions()
-      result = await driveFunc.deleteDrive(id)
-      print(result)
-      if result:
-        self.write({"message":"drive deleted successfully"})
-      else:
-        self.set_status(500)
-        self.write({"message":"database error"})
-        self.finish()
-    else:
-      self.set_status(400)
-      self.write({"message":"missing data"})
-      self.finish()
-
-  async def put(self, id):
-    if id:
       if self.request.body:
         data = json.loads(self.request.body)
-        driveFunc = DriveFunctions()
-        result = await driveFunc.updateDrive(id, data)
-        if result:
-          self.write({"message":"updated"})
+        if 'userId' in data:
+          driveFunc = DriveFunctions()
+          result = await driveFunc.getDrive(id)
+          if result:
+            if result['userId'] == data['userId']:
+              result = await driveFunc.deleteDrive(id)
+              if result:
+                self.write({"message":"drive deleted successfully"})
+              else:
+                self.set_status(500)
+                self.write({"message":"database error"})
+                self.finish()
+            else:
+              self.set_status(400)
+              self.write({"message":"wrong userId"})
+              self.finish()
+          else:
+            self.set_status(400)
+            self.write({"message":"drive not found"})
+            self.finish()
         else:
-          self.set_status(500)
-          self.write({"message":"database error"})
+          self.set_status(400)
+          self.write({"message":"missing userId"})
           self.finish()
       else:
         self.set_status(400)
@@ -77,5 +86,47 @@ class DrivesHandler(BaseHandler):
         self.finish()
     else:
       self.set_status(400)
-      self.write({"message":"missing data"})
+      self.write({"message":"missing drive id"})
       self.finish()
+
+  async def put(self, id):
+    if id:
+      if self.request.body:
+        data = json.loads(self.request.body)
+        if data['userId']:
+          driveFunc = DriveFunctions()
+          # making sure of the userId so not anyone update the drive but its owner
+          result = await driveFunc.getDrive(id)
+          # if there is no result then the drive doesn't exist
+          if result:
+            if result['userId'] == data['userId']:
+              result = await driveFunc.updateDrive(id, data)
+              if result:
+                self.write({"message":"updated"})
+              else:
+                self.set_status(500)
+                self.write({"message":"database error"})
+                self.finish()
+            else:
+              self.set_status(400)
+              self.write({"message":"wrong userId"})
+              self.finish()
+          else:
+            self.set_status(400)
+            self.write({"message":"drive not found"})
+            self.finish()
+        else:
+          self.set_status(400)
+          self.write({"message":"missing userId"})
+          self.finish()
+      else:
+        self.set_status(400)
+        self.write({"message":"missing data"})
+        self.finish()
+    else:
+      self.set_status(400)
+      self.write({"message":"missing drive id"})
+      self.finish()
+
+  def verify_data(self, data):
+    return data['userId'] and data['to'] and data['from'] and data['rideStatus'] and data['date'] and data['time'] and data['carId'] and data['proposal']
